fix(v2.5-hotfix): ProcessNameUtil 全版本进程名探测 + URL/apiKey 污染归一化 + 后端指纹加固
- 新增 util/ProcessNameUtil：读 /proc/self/cmdline 取进程名（Process.myProcessName 为 API 33+，
  minSdk=24 下 Android 7~12L 抛 NoSuchMethodError，为 v2.5 冷启动闪退根因）
- 新增 config/ApiUrl.normalizeBaseUrl：修复粘贴 URL 尾部 scheme 残渣（如 ...completionshttps）、
  自动补 https:// 前缀，幂等；model/apiKey 统一 trim
- 后端健康/认证指纹与诊断文本加固；CI 与测试同步更新

+++ app/src/test/java/com/chatbyyourside/llm/ModelAdmissionControllerTest.kt
@@ -140,7 +140,7 @@ class ModelAdmissionControllerTest {
     fun minContextExactlyFitsIsDowngradedToMinimum() {
         // 5.2GB 权重：仅 512 档（5.2G + 256M KV + 192M 预留 ≈ 5.64GB）放得进 5.75GB 预算 -> 降档到 512。
         val d = ModelAdmissionController.decideMemory(
-            mem(weightWorkingSet = 5.2f * gb.toFloat().toLong(), context = 4096, availMem = 6L * gb),
+            mem(weightWorkingSet = (5.2f * gb).toLong(), context = 4096, availMem = 6L * gb),
         ) as AdmissionDecision.Downgraded
         assertEquals(512, d.actualContext)
         assertTrue(d.reasons.contains(DowngradeReason.MEMORY))
